refactor and hide theme switcher for now
