+++ ui/src/routes/__root.tsx
@@ -2,7 +2,6 @@ import { createRootRoute, Outlet } from "@tanstack/react-router";
 import { lazy } from "react";
 
 import { TooltipProvider } from "@flow/components";
-import { ThemeProvider } from "@flow/components/ThemeProvider";
 import { config } from "@flow/config";
 import AuthenticationWrapper from "@flow/features/AuthenticationWrapper";
 import NotFoundPage from "@flow/features/NotFound";
@@ -10,6 +9,7 @@ import { NotificationSystem } from "@flow/features/NotificationSystem";
 import { AuthProvider } from "@flow/lib/auth";
 import { GraphQLProvider } from "@flow/lib/gql";
 import { I18nProvider } from "@flow/lib/i18n";
+import { ThemeProvider } from "@flow/lib/theme";
 
 export const Route = createRootRoute({
   component: RootRoute,
